nakanecto

+++ Data.Common/Services/ProductDbServices/IProductStateDbService.cs
@@ -3,7 +3,7 @@ using StepEbay.Data.Models.Products;
 
 namespace StepEbay.Data.Common.Services.ProductDbServices
 {
-    public interface IProductStateDbService : IDefaultDbService<byte, ProductState>
+    public interface IProductStateDbService : IDefaultDbService<int, ProductState>
     {
         public Task<bool> AnyStateByName(string name);
     }
